cli: Promote Agent Trace sync to a top-level command
Replace the former `sce trace` command group with the user-facing `sce sync` command. Move sync implementation and rendering into the sync service, remove trace inspection surfaces, and update CLI documentation and context contracts to reflect the new command surface.
