+++ context/cli/agent-trace-storage.md
@@ -22,10 +22,10 @@ Module at `cli/src/services/agent_trace_storage/` (T04 of the `repository-scoped
 
 ## Never-touch on-disk boundary
 
-The resolver never selects, creates, or touches pre-migration checkout-scoped `<state_root>/sce/agent-trace-<checkout-id>.db` files or the pre-migration global `<state_root>/sce/agent-trace.db`; tests assert neither appears after resolution. Active hook/runtime, lifecycle setup call sites, and `sce trace` status/shell resolution use this resolver. There is no longer a checkout-scoped resolver: the `retire-legacy-agent-trace-db` plan removed `checkout::resolve_or_create_agent_trace_db_for_checkout` and the `sce trace --legacy` inspection surface. Any pre-migration checkout/global DB files left on disk are never migrated, imported, renamed, or deleted, and are no longer inspectable through the CLI.
+The resolver never selects, creates, or touches pre-migration checkout-scoped `<state_root>/sce/agent-trace-<checkout-id>.db` files or the pre-migration global `<state_root>/sce/agent-trace.db`; tests assert neither appears after resolution. Active hook/runtime, lifecycle setup, and `sce sync` use this resolver. There is no longer a checkout-scoped resolver: the `retire-legacy-agent-trace-db` plan removed `checkout::resolve_or_create_agent_trace_db_for_checkout` and the former trace inspection surface. Any pre-migration checkout/global DB files left on disk are never migrated, imported, renamed, or deleted, and are no longer inspectable through the CLI.
 
 ## Status
 
-Registered in `cli/src/services/mod.rs` and consumed by hook runtime, Agent Trace lifecycle setup, and current-repository trace status/shell flows. T05 changed the resolved DB handle to the repository-scoped adapter and validates the stored `repository_metadata.repository_id` before returning storage; T08 wired hooks/lifecycle to pass resolved config values into this context; T09 wired trace UX to repository-scoped storage (the checkout-scoped `sce trace --legacy` surface was later removed by the `retire-legacy-agent-trace-db` plan). The `agent-trace-source-instance-id` plan's T03 split hook-runtime resolution into its own no-migration entrypoint, switching `open_agent_trace_db_for_hook_runtime` in `cli/src/services/hooks/mod.rs` off the setup/lifecycle resolver so hook runtime never runs migration `002` (or any migration). Covered by in-module tests: repository separation, SSH/HTTPS clone consolidation, linked-worktree consolidation, explicit-ID override, idempotent re-resolution, missing-identity guidance, path-segment validation, pre-migration checkout DB byte preservation/non-selection, empty fresh repository DB state, repository-level row sharing across equivalent clone checkouts, credential-safe remote canonicalization, concurrent first-open convergence, and hook-runtime resolution (fails before setup on a missing DB, fails before setup on a baseline-only pre-`002` schema without recording migration `002`, and matches setup's `RepositoryMetadata` once setup has run) (`nix build .#checks.<system>.cli-tests`).
+Registered in `cli/src/services/mod.rs` and consumed by hook runtime, Agent Trace lifecycle setup, and `sce sync`. T05 changed the resolved DB handle to the repository-scoped adapter and validates the stored `repository_metadata.repository_id` before returning storage; T08 wired hooks/lifecycle to pass resolved config values into this context; the former trace UX was later removed by the `retire-legacy-agent-trace-db` plan. The `agent-trace-source-instance-id` plan's T03 split hook-runtime resolution into its own no-migration entrypoint, switching `open_agent_trace_db_for_hook_runtime` in `cli/src/services/hooks/mod.rs` off the setup/lifecycle resolver so hook runtime never runs migration `002` (or any migration). Covered by in-module tests: repository separation, SSH/HTTPS clone consolidation, linked-worktree consolidation, explicit-ID override, idempotent re-resolution, missing-identity guidance, path-segment validation, pre-migration checkout DB byte preservation/non-selection, empty fresh repository DB state, repository-level row sharing across equivalent clone checkouts, credential-safe remote canonicalization, concurrent first-open convergence, and hook-runtime resolution (fails before setup on a missing DB, fails before setup on a baseline-only pre-`002` schema without recording migration `002`, and matches setup's `RepositoryMetadata` once setup has run) (`nix build .#checks.<system>.cli-tests`).
 
 See also: [repository-identity.md](repository-identity.md), [checkout-identity.md](checkout-identity.md), [default-path-catalog.md](default-path-catalog.md), [../sce/agent-trace-db.md](../sce/agent-trace-db.md)
