+++ context/cli/agent-trace-sync-command.md
@@ -1,16 +1,16 @@
 # Agent Trace sync architecture
 
-`sce trace sync` is the composition step that synchronizes a repository's local Agent Trace capture database with the control-plane Agent Trace ingestion API. It composes already-shipped infrastructure — repository/source identity, the read-only export readers, and the existing WorkOS auth/token-storage stack — into one command; it does not redesign the local database, source identity, export readers, or control-plane storage model.
+`sce sync` is the composition step that synchronizes a repository's local Agent Trace capture database with the control-plane Agent Trace ingestion API. It composes already-shipped infrastructure — repository/source identity, the read-only export readers, and the existing WorkOS auth/token-storage stack — into one command; it does not redesign the local database, source identity, export readers, or control-plane storage model.
 
 ## User flow
 
 ```
 sce auth login          # obtain and store WorkOS credentials
 cd <repository>          # any directory inside the target Git repository
-sce trace sync            # synchronize this repository's Agent Trace DB
+sce sync                  # synchronize this repository's Agent Trace DB
 ```
 
-`sce trace sync --format json` produces the same synchronization with machine-readable output; see [trace-command.md](trace-command.md) for the exact rendering contracts.
+`sce sync --format json` produces the same synchronization with machine-readable output; see [sync-command.md](sync-command.md) for the exact rendering contracts.
 
 ## Composed data flow
 
@@ -18,13 +18,13 @@ sce trace sync            # synchronize this repository's Agent Trace DB
 flowchart LR
     A[hooks / plugins] --> B[repository Agent Trace DB]
     B --> C[AgentTraceExportReader]
-    C --> D[sce trace sync]
+    C --> D[sce sync]
     D -- "HTTPS + WorkOS Bearer" --> E[control plane<br/>sce.crocoderlab.dev by default]
 ```
 
 - **hooks/plugins** write local capture rows (`messages`, `parts`, `diff_traces`, `agent_traces`) into the current repository's `RepositoryAgentTraceDb` during normal Git/editor activity — this is unchanged by sync.
 - **`AgentTraceExportReader`** (PR #198) is the read-only local export boundary sync uses to read rows after a cursor; sync never queries the repository DB directly.
-- **`sce trace sync`** resolves repository storage through the same `agent_trace_storage` path `sce trace status` uses (not the hook-runtime resolver), builds an `AuthenticatedControlPlaneClient` from stored WorkOS credentials and the resolved `control_plane_base_url`, and drives one authoritative `/state` call before starting four concurrent stream state machines. Each stream keeps its own batches and reconciliation refreshes sequential and cursor-safe; the bounded per-stream reconciliation loop remains independent. With no environment or config override, that base is `https://sce.crocoderlab.dev`; it is distinct from the `https://sce.crocoder.dev` SCE web and config-schema URL owner.
+- **`sce sync`** resolves repository storage through `agent_trace_storage` (not the hook-runtime resolver), builds an `AuthenticatedControlPlaneClient` from stored WorkOS credentials and the resolved `control_plane_base_url`, and drives one authoritative `/state` call before starting four concurrent stream state machines. Each stream keeps its own batches and reconciliation refreshes sequential and cursor-safe; the bounded per-stream reconciliation loop remains independent. With no environment or config override, that base is `https://sce.crocoderlab.dev`; it is distinct from the `https://sce.crocoder.dev` SCE web and config-schema URL owner.
 - **Credential runtime boundary:** `AuthenticatedControlPlaneClient` keeps the synchronous `CredentialStore` behind an `Arc` and runs every token-storage `load`/`save` through `tokio::task::spawn_blocking`. The underlying encrypted auth DB and Linux Secret Service/zbus APIs are blocking and may create their own Tokio runtime, so they must never execute directly inside the async control-plane request future. Token refresh and HTTP requests remain asynchronous; only credential persistence crosses the blocking boundary. The client owns a refresh single-flight guard: expired-token callers re-check credentials after acquiring it, and callers retrying the same rejected access token reuse a token saved by an earlier refresh; valid-token resolution does not acquire the guard.
 - **control plane** is the sole source of cursor truth: every invocation starts from `POST /agent-trace/ingestion/state`, uploads via `POST /agent-trace/ingestion/batch`, and advances a stream's cursor only from a validated batch response (`accepted == rows.len()` and `cursor == rows.last().sourceRowId`), never by inferring `cursor + rows.len()`.
 
@@ -39,7 +39,7 @@ Sync creates no local sync state anywhere on disk:
 - No Turso Sync and no direct Turso credentials in SCE.
 - No `BridgeLock` or local data-warehouse (DWH).
 
-Because every invocation starts from the control plane's authoritative `/state` cursors instead of local progress, restarts, conflicts, and ambiguous network failures are all recoverable without any client-side persisted state, and running `sce trace sync` twice in a row is naturally incremental — the second run's `/state` reflects the first run's uploads and only unsynced rows are re-read.
+Because every invocation starts from the control plane's authoritative `/state` cursors instead of local progress, restarts, conflicts, and ambiguous network failures are all recoverable without any client-side persisted state, and running `sce sync` twice in a row is naturally incremental — the second run's `/state` reflects the first run's uploads and only unsynced rows are re-read.
 
 ## Recovery semantics
 
@@ -54,8 +54,8 @@ Because every invocation starts from the control plane's authoritative `/state` 
 
 ## Related context
 
-- [trace-command.md](trace-command.md) — the full `sce trace` command group, including `sync` request/response rendering (text and JSON shapes).
-- [agent-trace-storage.md](agent-trace-storage.md) — the repository-scoped storage resolver sync reuses from `sce trace status`.
+- [sync-command.md](sync-command.md) — the `sce sync` command and its text/JSON rendering contract.
+- [agent-trace-storage.md](agent-trace-storage.md) — the repository-scoped storage resolver used by sync.
 - [agent-trace-export-readers.md](../sce/agent-trace-export-readers.md) — the read-only local export boundary sync reads through.
 - [auth-db.md](../sce/auth-db.md) — encrypted WorkOS credential storage sync authenticates through.
 - [Trace-sync progress stream contract](../decisions/2026-08-13-trace-sync-progress-stream-contract.md) — stderr progress/timestamps and stdout/JSON compatibility boundary.
