+++ context/sce/cli-shell-completion-contract.md
@@ -28,7 +28,7 @@ Defines the implemented `sce completion` contract for deterministic shell comple
 - Output is a shell script payload emitted on stdout for redirection/eval.
 - Scripts are deterministic for identical binary + input shell.
 - Generated scripts encode current parser-valid command/flag/subcommand surfaces for:
-- top-level commands: `help`, `config`, `setup`, `doctor`, `auth`, `hooks`, `trace`, `sync`, `version`, `completion`
+- top-level commands: `help`, `config`, `setup`, `doctor`, `auth`, `hooks`, `sync`, `version`, `completion`
   - completion-specific flags and values: `--shell` with `bash|zsh|fish`
 
 ## Implementation ownership
